+++ dao/dao.go
@@ -155,3 +155,16 @@ type Point struct {
 	Time  time.Time `json:"time"`
 	Price float64   `json:"price"`
 }
+
+type Delegators struct {
+	Items    []Delegator `json:"items"`
+	TotalCnt int         `json:"total_cnt"`
+	Offset   int         `json:"offset"`
+	Limit    int         `json:"limit"`
+}
+
+type Delegator struct {
+	Address    string `json:"address"`
+	ENSName    string `json:"ens_name"`
+	TokenValue string `json:"token_value"`
+}